style: move SidebarEntity
