+++ wondrous-app/components/Common/Sidebar/index.tsx
@@ -1,2 +1,1 @@
 export { default as SideBarComponent } from './Main';
-export { default as EntitySidebar } from './Entity';
