+++ wondrous-app/components/Common/SidebarEntity/AboutEntity/SettingsButton.tsx
