+++ wondrous-app/pages/organization/[username]/analytics.tsx
@@ -6,7 +6,7 @@ import { GET_USER_PERMISSION_CONTEXT } from 'graphql/queries';
 import { OrgBoardContext } from 'utils/contexts';
 import Analytics from 'components/organization/analytics';
 import { useGetOrgFromUsername } from 'utils/hooks';
-import { EntitySidebar } from 'components/Common/Sidebar';
+import EntitySidebar from 'components/Common/SidebarEntity';
 
 function ActivitiesPage() {
   const router = useRouter();
