+++ wondrous-app/pages/pod/[podId]/activities.tsx
@@ -6,7 +6,7 @@ import Activities from 'components/Pod/activities';
 import { GET_USER_PERMISSION_CONTEXT } from 'graphql/queries';
 import { PodBoardContext } from 'utils/contexts';
 import { useGetPodById } from 'utils/hooks';
-import { EntitySidebar } from 'components/Common/Sidebar';
+import EntitySidebar from 'components/Common/SidebarEntity';
 
 function ActivitiesPage() {
   const router = useRouter();
