+++ wondrous-app/pages/pod/[podId]/analytics.tsx
@@ -6,7 +6,7 @@ import { GET_USER_PERMISSION_CONTEXT } from 'graphql/queries';
 import { PodBoardContext } from 'utils/contexts';
 import Analytics from 'components/Pod/analytics';
 import { useGetPodById } from 'utils/hooks';
-import { EntitySidebar } from 'components/Common/Sidebar';
+import EntitySidebar from 'components/Common/SidebarEntity';
 
 function AnalyticsPage() {
   const router = useRouter();
